requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
pywebview>=4.0
Pillow>=10.0.0
